# refactor(crypto): Extract key generation into repo/gen. Speeds up tests.