--- a/cmd/cmd.go
+++ b/cmd/cmd.go
@@ -19,6 +19,7 @@ import (
 	"github.com/qri-io/ioes"
 	"github.com/qri-io/qri/lib"
 	"github.com/qri-io/qri/repo"
+	"github.com/qri-io/qri/repo/gen"
 )
 
 var log = golog.Logger("cmd")
@@ -41,7 +42,7 @@ func Execute() {
 
 	ensureLargeNumOpenFiles()
 
-	root := NewQriCommand(EnvPathFactory, ioes.NewStdIOStreams())
+	root := NewQriCommand(EnvPathFactory, gen.NewCryptoSource(), ioes.NewStdIOStreams())
 	// If the subcommand hits an error, don't show usage or the error, since we'll show
 	// the error message below, on our own. Usage is still shown if the subcommand
 	// is missing command-line arguments.