--- a/cmd/generate/generate.go
+++ b/cmd/generate/generate.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/qri-io/ioes"
 	"github.com/qri-io/qri/cmd"
+	"github.com/qri-io/qri/repo/gen"
 	"github.com/spf13/cobra/doc"
 )
 
@@ -17,7 +18,7 @@ func main() {
 	switch lastArg {
 	case "completions":
 		fmt.Printf("generating completions file...")
-		root := cmd.NewQriCommand(cmd.EnvPathFactory, ioes.NewStdIOStreams())
+		root := cmd.NewQriCommand(cmd.EnvPathFactory, gen.NewCryptoSource(), ioes.NewStdIOStreams())
 		root.GenBashCompletionFile("out.sh")
 		fmt.Println("done")
 	case "docs":
@@ -26,7 +27,7 @@ func main() {
 		if err := os.MkdirAll(path, os.ModePerm); err != nil {
 			log.Fatal(err)
 		}
-		root := cmd.NewQriCommand(cmd.EnvPathFactory, ioes.NewStdIOStreams())
+		root := cmd.NewQriCommand(cmd.EnvPathFactory, gen.NewCryptoSource(), ioes.NewStdIOStreams())
 		err := doc.GenMarkdownTree(root, path)
 		if err != nil {
 			log.Fatal(err)