@dustmop 2018-10-02
All comments addresses, pushing once tests complete...